feat: 유저 정보, 게시글 연결
