+++ src/components/Post/AddPost/index.tsx
@@ -1,54 +1,35 @@
-import React, { useState } from 'react'
-import styled from 'styled-components';
-import { supabase } from '../../../api/supabase';
+import React from 'react'
+import { useQuery, useQueryClient, useMutation } from '@tanstack/react-query'
+import { addPost } from '../../../api/post'
 
 const AddPost = () => {
-    type POST = {
-        id?: UUID;
-        userid?: string;
-        content?: string;
-        photo?: string;
-        isEditing?: boolean;
-        editingText?: string;
-        timeStamp?: number;
-    };
-    
-    type NewPost = Omit<POST, 'id' | 'createdAt' | 'editingText'>;
-
-    const [content, setContent] = useState('');
-
-    const handleContentInputchange = (e) => {
-        setContent(e.target.value)
+  const queryClient = useQueryClient();
+  
+  const addMutation = useMutation({
+    mutationFn: addPost,
+    onSuccess: () => {
+      queryClient.invalidateQueries({queryKey: ['posts']})
     }
+  })
 
-    // const handleAddPostBtnClick = 
-
-    const handleAddPostBtnClick = async (newPost: NewPost) => {
-        await supabase
-        .from('posts')
-        .insert(newPost);
-
-        console.log('newPost', newPost)
+  const handleChangeAddPost: React.ChangeEventHandler<HTMLInputElement> = (e) => {
+    e.preventDefault();
+    setContent(e.target.value);
+  }
+  
+  const handleSubmitAddPost: React.FormEventHandler<HTMLFormElement> = (e) => {
+    e.preventDefault();
+    const newPost = {
+      userid: 'test',
+      photo_url: 'posts?.photo_url',
+      content: content,
     };
-
+    addMutation.mutate(newPost);
+    setContent('');
+  };
   return (
-    <>
-        <StForm onSubmit={handleAddPostBtnClick}>
-            <input 
-                type='text'
-                placeholder='당신의 이야기를 공유해주세요'                
-                value={content}
-                name='content'
-                onChange={handleContentInputchange}
-            />
-            <button type='submit'>추가하기</button>
-        </StForm>
-    </>
+    <div>index</div>
   )
 }
 
 export default AddPost
-
-const StForm = styled.form`
-    margin-top: 100px;
-`
